rename file, remove testing code for uploading
